requirements.txt
numpy==1.18.2
pyserial==3.5b0
